add docstrings, move stowing logic with rest of manipulation logic
